Remove Ember cases, increase pings to 10,000 (still 1,000 at a time)
Removing Ember helps us focus on the thing we're trying to understand,
which is why ZIO 1 and Blaze work well but why calls hang with ZIO 2 and
Blaze.

The extra number of runs virtually guarantees the issue is replicated
in a given run of sbt zio-2/run.

Also add sbt-explicit-dependencies and update all.